# docs/swagger/swagger.yaml
basePath: /v1
definitions:
...
  dto.CreateSubscriptionRequest:
    properties:
      billing_cadence:
        $ref: '#/definitions/types.BillingCadence'
      billing_period:
        $ref: '#/definitions/types.BillingPeriod'
      billing_period_count:
        minimum: 1
        type: integer
      currency:
        type: string
      customer_id:
        type: string
      end_date:
        type: string
      lookup_key:
        type: string
      metadata:
        additionalProperties:
          type: string
        type: object
      plan_id:
        type: string
      start_date:
        type: string
      trial_end:
        type: string
      trial_start:
        type: string
    required:
    - billing_cadence
    - billing_period
    - billing_period_count
    - currency
    - customer_id
    - plan_id
    - start_date
    type: object
...
  dto.CustomerResponse:
    properties:
      address_city:
        description: AddressCity is the city of the customer's address
        type: string
      address_country:
        description: AddressCountry is the country of the customer's address (ISO
          3166-1 alpha-2)
        type: string
      address_line1:
        description: AddressLine1 is the first line of the customer's address
        type: string
      address_line2:
        description: AddressLine2 is the second line of the customer's address
        type: string
      address_postal_code:
        description: AddressPostalCode is the postal code of the customer's address
        type: string
      address_state:
        description: AddressState is the state of the customer's address
        type: string
      created_at:
        type: string
      created_by:
        type: string
      email:
        description: Email is the email of the customer
        type: string
      environment_id:
        description: EnvironmentID is the environment identifier for the customer
        type: string
      external_id:
        description: ExternalID is the external identifier for the customer
        type: string
      id:
        description: ID is the unique identifier for the customer
        type: string
      metadata:
        additionalProperties:
          type: string
        description: Metadata
        type: object
      name:
        description: Name is the name of the customer
        type: string
      status:
        $ref: '#/definitions/types.Status'
      tenant_id:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  dto.GetUsageBySubscriptionRequest:
    properties:
      end_time:
        example: "2024-03-20T00:00:00Z"
        type: string
      lifetime_usage:
        example: false
        type: boolean
      start_time:
        example: "2024-03-13T00:00:00Z"
        type: string
      subscription_id:
        example: "123"
        type: string
    required:
    - subscription_id
    type: object
  dto.GetUsageBySubscriptionResponse:
    properties:
      amount:
        type: number
      charges:
        items:
          $ref: '#/definitions/dto.SubscriptionUsageByMetersResponse'
        type: array
      currency:
        type: string
      display_amount:
        type: string
      end_time:
        type: string
      start_time:
        type: string
    type: object
...
  dto.ListSubscriptionPausesResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.SubscriptionPauseResponse'
        type: array
      total:
        type: integer
    type: object
  dto.ListSubscriptionsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.SubscriptionResponse'
        type: array
      pagination:
        $ref: '#/definitions/types.PaginationResponse'
    type: object
...
  dto.PauseSubscriptionRequest:
    properties:
      dry_run:
        type: boolean
      metadata:
        additionalProperties:
          type: string
        type: object
      pause_days:
        type: integer
      pause_end:
        type: string
      pause_mode:
        $ref: '#/definitions/types.PauseMode'
      pause_start:
        type: string
      reason:
        maxLength: 255
        type: string
    required:
    - pause_mode
    type: object
...
  dto.ResumeSubscriptionRequest:
    properties:
      dry_run:
        type: boolean
      metadata:
        additionalProperties:
          type: string
        type: object
      resume_mode:
        $ref: '#/definitions/types.ResumeMode'
    required:
    - resume_mode
    type: object
...
  dto.SubscriptionPauseResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      environment_id:
        description: EnvironmentID is the environment identifier for the pause
        type: string
      id:
        description: ID is the unique identifier for the subscription pause
        type: string
      metadata:
        allOf:
        - $ref: '#/definitions/types.Metadata'
        description: Metadata is a map of key-value pairs that can be attached to
          the pause
      original_period_end:
        description: OriginalPeriodEnd is the end of the billing period when the pause
          was created
        type: string
      original_period_start:
        description: OriginalPeriodStart is the start of the billing period when the
          pause was created
        type: string
      pause_end:
        description: PauseEnd is when the pause will end (null for indefinite)
        type: string
      pause_mode:
        allOf:
        - $ref: '#/definitions/types.PauseMode'
        description: PauseMode indicates how the pause was applied
      pause_start:
        description: PauseStart is when the pause actually started
        type: string
      pause_status:
        allOf:
        - $ref: '#/definitions/types.PauseStatus'
        description: PauseStatus is the status of the pause
      reason:
        description: Reason is the reason for pausing
        type: string
      resume_mode:
        allOf:
        - $ref: '#/definitions/types.ResumeMode'
        description: ResumeMode indicates how the resume will be applied
      resumed_at:
        description: ResumedAt is when the pause was actually ended (if manually resumed)
        type: string
      status:
        $ref: '#/definitions/types.Status'
      subscription_id:
        description: SubscriptionID is the identifier for the subscription
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  dto.SubscriptionResponse:
    properties:
      active_pause_id:
        description: |-
          ActivePauseID references the current active pause configuration
          This will be null if no pause is active or scheduled
        type: string
      billing_anchor:
        description: |-
          BillingAnchor is the reference point that aligns future billing cycle dates.
          It sets the day of week for week intervals, the day of month for month and year intervals,
          and the month of year for year intervals. The timestamp is in UTC format.
        type: string
      billing_cadence:
        allOf:
        - $ref: '#/definitions/types.BillingCadence'
        description: BillingCadence is the cadence of the billing cycle.
      billing_period:
        allOf:
        - $ref: '#/definitions/types.BillingPeriod'
        description: BillingPeriod is the period of the billing cycle.
      billing_period_count:
        description: BillingPeriodCount is the total number units of the billing period.
        type: integer
      cancel_at:
        description: CancelAt is the date the subscription will be canceled
        type: string
      cancel_at_period_end:
        description: CancelAtPeriodEnd is whether the subscription was canceled at
          the end of the current period
        type: boolean
      cancelled_at:
        description: CanceledAt is the date the subscription was canceled
        type: string
      created_at:
        type: string
      created_by:
        type: string
      currency:
        description: Currency is the currency of the subscription in lowercase 3 digit
          ISO codes
        type: string
      current_period_end:
        description: |-
          CurrentPeriodEnd is the end of the current period that the subscription has been invoiced for.
          At the end of this period, a new invoice will be created.
        type: string
      current_period_start:
        description: |-
          CurrentPeriodStart is the end of the current period that the subscription has been invoiced for.
          At the end of this period, a new invoice will be created.
        type: string
      customer:
        $ref: '#/definitions/dto.CustomerResponse'
      customer_id:
        description: CustomerID is the identifier for the customer in our system
        type: string
      end_date:
        description: EndDate is the end date of the subscription
        type: string
      environment_id:
        description: EnvironmentID is the environment identifier for the subscription
        type: string
      id:
        description: ID is the unique identifier for the subscription
        type: string
      line_items:
        items:
          $ref: '#/definitions/subscription.SubscriptionLineItem'
        type: array
      lookup_key:
        description: LookupKey is the key used to lookup the subscription in our system
        type: string
      metadata:
        allOf:
        - $ref: '#/definitions/types.Metadata'
        description: Metadata is a map of key-value pairs that can be attached to
          the subscription
      pause_status:
        allOf:
        - $ref: '#/definitions/types.PauseStatus'
        description: PauseStatus tracks the current pause state
      pauses:
        items:
          $ref: '#/definitions/subscription.SubscriptionPause'
        type: array
      plan:
        $ref: '#/definitions/dto.PlanResponse'
      plan_id:
        description: PlanID is the identifier for the plan in our system
        type: string
      start_date:
        description: StartDate is the start date of the subscription
        type: string
      status:
        $ref: '#/definitions/types.Status'
      subscription_status:
        allOf:
        - $ref: '#/definitions/types.SubscriptionStatus'
        description: Status is the status of the subscription
      tenant_id:
        type: string
      trial_end:
        description: TrialEnd is the end date of the trial period
        type: string
      trial_start:
        description: TrialStart is the start date of the trial period
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      version:
        description: Version is used for optimistic locking
        type: integer
    type: object
...
  errors.ErrorDetail:
    properties:
      details:
        additionalProperties: {}
        type: object
      message:
        type: string
    type: object
  errors.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/errors.ErrorDetail'
      success:
        type: boolean
    type: object
  gin.H:
    additionalProperties: {}
    type: object
...
  subscription.SubscriptionLineItem:
    properties:
      billing_period:
        $ref: '#/definitions/types.BillingPeriod'
      created_at:
        type: string
      created_by:
        type: string
      currency:
        type: string
      customer_id:
        type: string
      display_name:
        type: string
      end_date:
        type: string
      environment_id:
        type: string
      id:
        type: string
      invoice_cadence:
        $ref: '#/definitions/types.InvoiceCadence'
      metadata:
        additionalProperties:
          type: string
        type: object
      meter_display_name:
        type: string
      meter_id:
        type: string
      plan_display_name:
        type: string
      plan_id:
        type: string
      price_id:
        type: string
      price_type:
        $ref: '#/definitions/types.PriceType'
      quantity:
        type: number
      start_date:
        type: string
      status:
        $ref: '#/definitions/types.Status'
      subscription_id:
        type: string
      tenant_id:
        type: string
      trial_period:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  subscription.SubscriptionPause:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      environment_id:
        description: EnvironmentID is the environment identifier for the pause
        type: string
      id:
        description: ID is the unique identifier for the subscription pause
        type: string
      metadata:
        allOf:
        - $ref: '#/definitions/types.Metadata'
        description: Metadata is a map of key-value pairs that can be attached to
          the pause
      original_period_end:
        description: OriginalPeriodEnd is the end of the billing period when the pause
          was created
        type: string
      original_period_start:
        description: OriginalPeriodStart is the start of the billing period when the
          pause was created
        type: string
      pause_end:
        description: PauseEnd is when the pause will end (null for indefinite)
        type: string
      pause_mode:
        allOf:
        - $ref: '#/definitions/types.PauseMode'
        description: PauseMode indicates how the pause was applied
      pause_start:
        description: PauseStart is when the pause actually started
        type: string
      pause_status:
        allOf:
        - $ref: '#/definitions/types.PauseStatus'
        description: PauseStatus is the status of the pause
      reason:
        description: Reason is the reason for pausing
        type: string
      resume_mode:
        allOf:
        - $ref: '#/definitions/types.ResumeMode'
        description: ResumeMode indicates how the resume will be applied
      resumed_at:
        description: ResumedAt is when the pause was actually ended (if manually resumed)
        type: string
      status:
        $ref: '#/definitions/types.Status'
      subscription_id:
        description: SubscriptionID is the identifier for the subscription
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  types.BillingCadence:
    enum:
    - RECURRING
    - ONETIME
    type: string
    x-enum-varnames:
    - BILLING_CADENCE_RECURRING
    - BILLING_CADENCE_ONETIME
...
  types.BillingPeriod:
    enum:
    - MONTHLY
    - ANNUAL
    - WEEKLY
    - DAILY
    - QUARTERLY
    - HALF_YEARLY
    type: string
    x-enum-varnames:
    - BILLING_PERIOD_MONTHLY
    - BILLING_PERIOD_ANNUAL
    - BILLING_PERIOD_WEEKLY
    - BILLING_PERIOD_DAILY
    - BILLING_PERIOD_QUARTER
    - BILLING_PERIOD_HALF_YEAR
...
  types.Metadata:
    additionalProperties:
      type: string
    type: object
  types.PaginationResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  types.PauseMode:
    enum:
    - immediate
    - scheduled
    - period_end
    type: string
    x-enum-varnames:
    - PauseModeImmediate
    - PauseModeScheduled
    - PauseModePeriodEnd
  types.PauseStatus:
    enum:
    - none
    - active
    - scheduled
    - completed
    - cancelled
    type: string
    x-enum-varnames:
    - PauseStatusNone
    - PauseStatusActive
    - PauseStatusScheduled
    - PauseStatusCompleted
    - PauseStatusCancelled
...
  types.Status:
    enum:
    - published
    - deleted
    - archived
    type: string
    x-enum-varnames:
    - StatusPublished
    - StatusDeleted
    - StatusArchived
...
info:
  contact: {}
  description: FlexPrice API Service
  title: FlexPrice API
  version: "1.0"
paths:
...
  /subscriptions:
    get:
      description: Get subscriptions with optional filtering
      parameters:
      - description: ActiveAt filters subscriptions that are active at the given time
        in: query
        name: active_at
        type: string
      - collectionFormat: csv
        description: BillingCadence filters by billing cadence
        in: query
        items:
          enum:
          - RECURRING
          - ONETIME
          type: string
        name: billing_cadence
        type: array
      - collectionFormat: csv
        description: BillingPeriod filters by billing period
        in: query
        items:
          enum:
          - MONTHLY
          - ANNUAL
          - WEEKLY
          - DAILY
          - QUARTERLY
          - HALF_YEARLY
          type: string
        name: billing_period
        type: array
      - description: CustomerID filters by customer ID
        in: query
        name: customer_id
        type: string
      - in: query
        name: end_time
        type: string
      - in: query
        name: expand
        type: string
      - description: IncludeCanceled includes canceled subscriptions if true
        in: query
        name: include_canceled
        type: boolean
      - in: query
        maximum: 1000
        minimum: 1
        name: limit
        type: integer
      - in: query
        minimum: 0
        name: offset
        type: integer
      - enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - description: PlanID filters by plan ID
        in: query
        name: plan_id
        type: string
      - in: query
        name: sort
        type: string
      - in: query
        name: start_time
        type: string
      - enum:
        - published
        - deleted
        - archived
        in: query
        name: status
        type: string
        x-enum-varnames:
        - StatusPublished
        - StatusDeleted
        - StatusArchived
      - collectionFormat: csv
        description: SubscriptionStatus filters by subscription status
        in: query
        items:
          enum:
          - active
          - paused
          - cancelled
          - incomplete
          - incomplete_expired
          - past_due
          - trialing
          - unpaid
          type: string
        name: subscription_status
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListSubscriptionsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List subscriptions
      tags:
      - Subscriptions
    post:
      consumes:
      - application/json
      description: Create a new subscription
      parameters:
      - description: Subscription Request
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create subscription
      tags:
      - Subscriptions
  /subscriptions/{id}:
    get:
      description: Get a subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get subscription
      tags:
      - Subscriptions
  /subscriptions/{id}/cancel:
    post:
      consumes:
      - application/json
      description: Cancel a subscription
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      - description: Cancel at period end
        in: query
        name: cancel_at_period_end
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Cancel subscription
      tags:
      - Subscriptions
  /subscriptions/{id}/pause:
    post:
      consumes:
      - application/json
      description: Pause a subscription with the specified parameters
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      - description: Pause subscription request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PauseSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SubscriptionPauseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Pause a subscription
      tags:
      - Subscriptions
  /subscriptions/{id}/pauses:
    get:
      description: List all pauses for a subscription
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ListSubscriptionPausesResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: List all pauses for a subscription
      tags:
      - Subscriptions
  /subscriptions/{id}/resume:
    post:
      consumes:
      - application/json
      description: Resume a paused subscription with the specified parameters
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      - description: Resume subscription request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ResumeSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SubscriptionPauseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Resume a paused subscription
      tags:
      - Subscriptions
  /subscriptions/usage:
    post:
      consumes:
      - application/json
      description: Get usage for a subscription
      parameters:
      - description: Usage request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.GetUsageBySubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetUsageBySubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get usage by subscription
      tags:
      - Subscriptions
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: Enter your API key in the format *x-api-key &lt;api-key&gt;**
    in: header
    name: x-api-key
    type: apiKey
swagger: "2.0"